pyproject.toml
[project]
name = "jamal"
version = "0.1.0"
description = "MP3 stem separation + chromagram visualization"
requires-python = ">=3.11"
dependencies = [
    "demucs>=4.0.0",
    "librosa>=0.10.0",
    "matplotlib>=3.8.0",
    "Pillow>=10.0.0",
    "rich>=13.0.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "numpy>=1.24.0",
]

[project.scripts]
jamal = "jamal.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
